Removed db safety check
This safety check was to watch out for another process using the same database instance.
It did so by assuming the value of maxGap would be the same in memory as in the database.
Unfortunately, the unit clears the database between tests and causes this false error.
Decided to remove this safety check and someday replace with a stronger one.